kv: Rearrange code in send.go
Some parts are only used by Transports, so they go in transport.go. The
send() function itself moves to dist_sender.go.

Pure cut-and-paste.
